requirements.txt
fastapi
uvicorn
jinja2
yt-dlp
openai-whisper
ffmpeg-python
transformers
torch